requirements.txt
Flask
Flask-SQLAlchemy
Flask-Script
gunicorn
werkzeug
